# PERF: Groupby.count