```
from asv_bench.benchmarks.groupby import *
self4 = GroupByMethods()
self4.setup('uint', 'count', 'direct', 5)

%timeit self4.time_dtype_as_field('uint', 'count', 'direct', 5)
231 µs ± 5.23 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- master
77.3 µs ± 619 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- PR
```